Move MathDomain to sphinx.domains.math
