+++ doc/extdev/index.rst
@@ -126,6 +126,11 @@ The following is a list of deprecated interface.
      - 4.0
      - :meth:`~sphinx.application.Sphinx.add_css_file()`
 
+   * - ``sphinx.ext.mathbase.MathDomain``
+     - 1.8
+     - 3.0
+     - ``sphinx.domains.math.MathDomain``
+
    * - ``sphinx.ext.mathbase.is_in_section_title()``
      - 1.8
      - 3.0
